backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Vitalii Shishorin <moskrc@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.6"
django-environ = "^0.4.5"
Keras = "^2.4.3"
tensorflow = "^2.4.1"
Pillow = "^8.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
